refactor: 프로젝트 구조 세분화(#14)
